requirements.txt
numpy>=1.24
pandas>=2.0
soundfile>=0.12
librosa>=0.10
h5py>=3.9
scikit-learn>=1.3
hdbscan>=0.8.33
umap-learn>=0.5
pyyaml>=6.0
tqdm>=4.65
wandb>=0.15
